# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zinora AI - AI Customer Support Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface SignupInput {
  name: string;
  email: string;
  password: string;
  company: string;
}

export interface ForgotPasswordInput {
  email: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  company: string;
  plan: string;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserUpdate {
  name?: string;
  company?: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface DashboardSummary {
  totalConversations: number;
  resolvedToday: number;
  activeChats: number;
  avgResponseTime: string;
  satisfactionScore: number;
  aiResolutionRate: number;
  weeklyGrowth: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  timestamp: string;
  /** @nullable */
  user?: string | null;
}

export interface Conversation {
  id: number;
  customerName: string;
  customerEmail: string;
  subject: string;
  status: string;
  lastMessage: string;
  lastMessageAt: string;
  tags: string[];
  channel: string;
  /** @nullable */
  assignedTo: string | null;
  unreadCount: number;
  createdAt: string;
}

export interface ConversationInput {
  customerName: string;
  customerEmail: string;
  subject: string;
  channel: string;
}

export interface ConversationUpdate {
  status?: string;
  tags?: string[];
  /** @nullable */
  assignedTo?: string | null;
}

export interface Message {
  id: number;
  conversationId: number;
  content: string;
  sender: string;
  senderType: string;
  timestamp: string;
  isAi: boolean;
  attachments?: string[];
}

export interface MessageInput {
  content: string;
  attachments?: string[];
}

export interface Article {
  id: number;
  title: string;
  content: string;
  category: string;
  tags: string[];
  status: string;
  views: number;
  createdAt: string;
  updatedAt: string;
}

export interface ArticleInput {
  title: string;
  content: string;
  category: string;
  tags?: string[];
  status?: string;
}

export interface ArticleUpdate {
  title?: string;
  content?: string;
  category?: string;
  tags?: string[];
  status?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  articleCount: number;
}

export interface ChartDataPoint {
  name: string;
  value: number;
}

export interface AnalyticsOverview {
  totalTickets: number;
  resolvedTickets: number;
  avgSatisfaction: number;
  avgResponseTime: number;
  aiResolutionRate: number;
  ticketsByChannel: ChartDataPoint[];
  ticketsByStatus: ChartDataPoint[];
}

export interface TimeSeriesPoint {
  date: string;
  value: number;
}

export interface TeamPerformancePoint {
  name: string;
  resolved: number;
  avgTime: number;
  satisfaction: number;
}

export interface AnalyticsCharts {
  dailyConversations: TimeSeriesPoint[];
  satisfactionTrend: TimeSeriesPoint[];
  responseTimeTrend: TimeSeriesPoint[];
  teamPerformance: TeamPerformancePoint[];
}

export interface TeamMember {
  id: number;
  name: string;
  email: string;
  role: string;
  status: string;
  joinedAt: string;
  conversationsHandled: number;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface TeamInviteInput {
  email: string;
  role: string;
  name?: string;
}

export interface TeamMemberUpdate {
  role?: string;
  status?: string;
}

export interface Subscription {
  planId: string;
  planName: string;
  status: string;
  currentPeriodStart: string;
  currentPeriodEnd: string;
  aiCreditsUsed: number;
  aiCreditsLimit: number;
  seatsUsed: number;
  seatsLimit: number;
}

export interface Plan {
  id: string;
  name: string;
  price: number;
  yearlyPrice: number;
  description: string;
  features: string[];
  isPopular: boolean;
  aiCredits: number;
  seats: number;
}

export interface Invoice {
  id: string;
  amount: number;
  status: string;
  date: string;
  description: string;
  downloadUrl: string;
}

export interface Notification {
  id: number;
  type: string;
  title: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface AiSettings {
  id: number;
  personality: string;
  responseTone: string;
  model: string;
  language: string;
  confidenceThreshold: number;
  autoEscalate: boolean;
  autoEscalateThreshold: number;
  maxResponseLength: number;
  enableSentimentAnalysis: boolean;
  enableAutoReply: boolean;
  workingHoursOnly: boolean;
}

export interface AiSettingsUpdate {
  personality?: string;
  responseTone?: string;
  model?: string;
  language?: string;
  confidenceThreshold?: number;
  autoEscalate?: boolean;
  autoEscalateThreshold?: number;
  maxResponseLength?: number;
  enableSentimentAnalysis?: boolean;
  enableAutoReply?: boolean;
  workingHoursOnly?: boolean;
}

export interface Integration {
  id: string;
  name: string;
  description: string;
  category: string;
  isConnected: boolean;
  logoUrl: string;
  /** @nullable */
  connectedAt?: string | null;
}

export interface ContactInput {
  name: string;
  email: string;
  company?: string;
  subject: string;
  message: string;
}

export type ListConversationsParams = {
status?: string;
search?: string;
tag?: string;
};

export type ListArticlesParams = {
category?: string;
search?: string;
};

export type GetAnalyticsOverviewParams = {
period?: string;
};

export type GetAnalyticsChartsParams = {
period?: string;
};