# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zinora AI - AI Customer Support Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  company: string;
  plan: string;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}