# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zinora AI - AI Customer Support Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  company?: string;
  /** @nullable */
  avatarUrl?: string | null;
}